ch(refactoredAPI):
Made changes to the file structure to accomodate modular design and object oriented design
[Fixes #160845226]
